[refactor] (bitcoincash) move common HD test to helper.ts
